Proposal: Extract semantics from list of log entries
Because of the fact, that our Engineers had concerns regarding too much
overhead in log files, we propose a version that minimizes the semantic
overhead without loosing semantic information.

It would be nice to allow this way of structuring the log messages
additional to our "classic" way of modelling events, that we need
(especially for the event publishing via MQTT).