Have less restrictions for gateway/events filtering.
Before this change, to get all events related to a specific port
(with origin PORT or DEVICE), the user has to do two requests.
With this change, a single
    GET /gateway/events?origin=ALL&masterNumber=1&portNumber=2
request will do the job (= reading all events from master 1 port 2
or from the device connected to this port).
This resolves #118.